requirements.txt
getgauge>=0.4.0
requests>=2.31
colorlog>=6.7
pydantic>=2.4.2
behave>=1.2.6
behavex>=3.0.0
requests-to-curl>=1.1
mysql-connector-python>=8.0.30
pymongo>=4.2.0
redis>=4.3.4
psycopg2-binary>=2.9.7
TestLink-API-Python-client>=0.8.1
locust>=2.16.1
python-gitlab>=3.15.0
webdriver_manager>=4.0.1
selenium==4.4.3
Appium-Python-Client==2.9.0
dataclasses-json>=0.6.2
Faker==19.6.2
jsonschema
